requirements.txt
dj-database-url==0.4.2
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-heroku==0.2.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Markdown==2.6.10
mccabe==0.6.1
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pytz==2017.3
whitenoise==3.3.1
